Fixed module resolutions
