+++ package.json
@@ -2,7 +2,19 @@
   "name": "@genshin-toolkit/parser",
   "version": "1.0.1",
   "description": "A small library to parse and validate Genshin Impact game data.",
-  "main": "dist/index.js",
+  "main": "dist/node/index.js",
+  "types": "dist/node/index.ts",
+  "exports": {
+    ".": {
+      "node": "./dist/node/index.js",
+      "default": "./dist/web/index.js"
+    },
+    "./web": {
+      "types": "./dist/web/web/index.d.ts",
+      "default": "./dist/web/web/index.js"
+    }
+  },
+  "files": ["dist"],
   "scripts": {
     "build:node": "tsc --project tsconfig.node.json",
     "build:web": "tsc --project tsconfig.web.json",
