+++ src/index.ts
@@ -1,4 +1,5 @@
-export { BufferDataProvider, DataProvider, FileDataProvider, StringDataProvider } from './lib/data-provider';
+export { BufferDataProvider, DataProvider, StringDataProvider } from './lib/data-provider';
+export { FileDataProvider } from './lib/data-provider-node';
 export { GameData } from './lib/gamedata-schema';
 export * as Har from './lib/har.module';
 export { MhyResponse } from './lib/mhyresponse.module';
