+++ src/web/data-provider.ts
@@ -1,4 +1,4 @@
-import { DataProvider } from '../../lib/data-provider';
+import { DataProvider } from '../lib/data-provider';
 
 /**
  * Data provider that loads data from a File object using FileReader API.
