+++ src/web/index.ts
@@ -1,6 +1,7 @@
+export { BufferDataProvider, DataProvider, StringDataProvider } from '../lib/data-provider';
 export { GameData } from '../lib/gamedata-schema';
 export * as Har from '../lib/har.module';
 export { MhyResponse } from '../lib/mhyresponse.module';
-export { DataProvider, BufferDataProvider, StringDataProvider } from '../lib/data-provider';
-export { WebFileDataProvider } from './lib/data-provider';
 export { GAME_RECORD_API_BASEURL, loadFromFile, loadFromHar } from '../lib/parser';
+export { WebFileDataProvider } from './data-provider';
+
