+++ tsconfig.node.json
@@ -1,8 +1,8 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "module": "commonjs",                                /* Specify what module code is generated. */
-    "outDir": "dist",                                    /* Specify an output folder for all emitted files. */
+    "outDir": "dist/node",
   },
-  "include": [ "src" ]
+  "include": ["src"],
+  "exclude": ["src/web"]  // Exclude web-specific code
 }
