+++ tsconfig.web.json
@@ -1,10 +1,8 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "lib": ["DOM", "ES2020"],                            /* Specify a set of bundled library declaration files that describe the target runtime environment. */
-    "module": "commonjs",                                /* Specify what module code is generated. */
-    "outDir": "dist/web",                                    /* Specify an output folder for all emitted files. */
+    "lib": ["DOM", "ES2020"],
+    "outDir": "dist/web"
   },
-  "exclude": ["src/tests"],
-  "include": [ "src" ]
+  "include": ["src/web"]
 }
